chore: move form-spatial component
