+++ src/components/form-spatial/configured-form-spatial.jsx
@@ -2,7 +2,7 @@ import { reduxForm } from 'redux-form';
 
 import FormSpatial from './form-spatial.component';
 import validate from './form-spatial-validations';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'spatial',
