+++ src/components/form-spatial/connected-form-spatial.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createMockStore } from 'redux-test-utils';
-import shallowWithStore from '../../../../test/shallowWithStore';
+import shallowWithStore from '../../../test/shallowWithStore';
 import { ConnectedFormSpatial } from './connected-form-spatial.component';
-import dataset from '../../../../test/fixtures/datasets';
+import dataset from '../../../test/fixtures/datasets';
 
 test('should render ConnectedFormSpatial correctly', () => {
   const testState = {};
